Fix number overflow

+++ sample.py
@@ -27,7 +27,7 @@ temperature = 0.8 # 1.0 = no change, < 1.0 = less random, > 1.0 = more random, i
 top_k = 200 # retain only the top_k most likely tokens, clamp others to have 0 probability
 seed = 42069
 device = torch.device("mps") # examples: 'cpu', 'cuda', 'cuda:0', 'cuda:1', etc.
-dtype = 'float16' # 'bfloat16' or 'float16'
+dtype = 'float32' # 'bfloat16' or 'float16'
 show_probs = False # Set to True to see chart of top 10 tokens each iteration
 compile = True # use PyTorch 2.0 to compile the model to be faster
 exec(open('configurator.py').read()) # overrides from command line or config file
